Refactor PaymentMethodRepository.
